+++ src/main/java/net/frozenblock/wilderwild/block/SeedingFlowerBlock.java
@@ -20,7 +20,9 @@ package net.frozenblock.wilderwild.block;
 
 import com.mojang.serialization.MapCodec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
+import net.frozenblock.wilderwild.config.WWBlockConfig;
 import net.frozenblock.wilderwild.particle.options.SeedParticleOptions;
+import net.frozenblock.wilderwild.registry.WWBlocks;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Holder;
 import net.minecraft.core.registries.BuiltInRegistries;
@@ -96,6 +98,7 @@ public class SeedingFlowerBlock extends FlowerBlock {
 	}
 
 	public boolean canShearIntoOriginalFlower(LevelReader world, BlockPos pos, BlockState state) {
+		if (this == WWBlocks.SEEDING_DANDELION) return WWBlockConfig.get().flower.shearSeedingDandelions;
 		return true;
 	}